refactor: extraer protocolo DeviceBridge y reorganizar CLI en AutoCore + AutoLibiOS
Prepara la arquitectura para soportar iOS y Android como binarios separados:

- Nuevo target AutoCore: DeviceBridge (protocolo 22 métodos), BridgeError,
  CommandDispatcher, ScriptParser, Config, TreePrinter
- Nuevo target AutoLibiOS: SimulatorBridge (conforma DeviceBridge),
  ElementIndex, TargetResolver, UIStabilizer, AXDebug, BuildInterceptor,
  DylibInjector, MockHeaders
- Nuevo binario stub auto-android (compila contra AutoCore, sin deps iOS)
- Elimina AutoLib (reemplazado por AutoCore + AutoLibiOS)
- Corrige acentos y redacción ambigua en docs/libro/06-alternativas.md
- Agrega exclusiones Android al .gitignore

27/27 tests pasan. Ambos binarios compilan.
