+++ cli/Sources/AutoCore/ScriptParser.swift
