+++ cli/Sources/AutoCore/TreePrinter.swift
