+++ cli/Sources/CLIAndroid/main.swift
@@ -1,0 +1,11 @@
+import Foundation
+import AutoCore
+
+print("auto-android: not yet implemented")
+print("This binary will support Android device automation via ADB.")
+print("\nShared infrastructure ready:")
+print("  - DeviceBridge protocol (\(22) methods)")
+print("  - CommandDispatcher (shared commands)")
+print("  - ScriptParser (.auto scripts)")
+print("  - Config (.autopilot)")
+exit(1)
