+++ cli/Tests/TokenizerTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-@testable import AutoLib
+@testable import AutoCore
 
 final class TokenizerTests: XCTestCase {
